Finance Tooltip UI connected, Currency UI updates now
Startfunction in GameManager

+++ RailroadTrader_UnityProject/Assets/Scripts/Buildings/SupplyStores.cs
@@ -19,7 +19,7 @@ public class SupplyStores : Building
         public Passanger type;
         public int maxCapacity;
         public int curAmount;
-        public float earningGain;
+        public int earningGain;
         public float satisfactionGain;
     }
